fix: address v2 review findings - decimal hash mismatch, missing product_type, sha256 consistency
1. Cast abl.quantity::float8 in v1.2.2 server-side hash SELECT so JSONB
   serializes without trailing zeros, matching JavaScript's representation.

2. Add productType field to BomOutputLine interface, propagate through all
   pipeline calculation paths (WALL_PANEL, LIGHT, FURNITURE, HIDDEN_COMPONENT),
   and include product_type in saveBomToServer RPC payload mapping.

3. Replace pg_catalog.sha256() with digest(..., 'sha256') in v1.1.8
   create_project for consistency with v1.2.0/v1.2.2.

+++ migrations/v1.2.2_finalization_integrity.sql
@@ -141,7 +141,7 @@ BEGIN
                         SELECT abl.actual_bom_line_id,
                                abl.sku_id,
                                abl.product_type,
-                               abl.quantity,
+                               abl.quantity::float8,
                                abl.required_quantity,
                                abl.waste_quantity,
                                abl.unit_of_measure,
